Add files via upload
This is a small macro that reads the metadata on the JLS lab 880.  It kinda works for the 980. Sometimes the fields are not completed correctly.  In general it just used the bioformats commands and hunts for the info in that giant metadata file. It's  a work in progress. Use and edit freely.